split test_special_ops
