+++ tests/test_special_ops.py
@@ -9,7 +9,6 @@ from .accuracy_utils import (
     FLOAT_DTYPES,
     POINTWISE_SHAPES,
     gems_assert_close,
-    gems_assert_equal,
     to_reference,
 )
 
@@ -196,59 +195,6 @@ def test_embedding(EmbeddingSize, Batch, M, N, padding_idx, scale_grad_by_freq, 
 
 
 @pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_rand(shape, dtype):
-    with flag_gems.use_gems():
-        res_out = torch.rand(shape, dtype=dtype, device="cuda")
-    assert (res_out <= 1.0).all()
-    assert (res_out >= 0.0).all()
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_randn(shape, dtype):
-    with flag_gems.use_gems():
-        res_out = torch.randn(shape, dtype=dtype, device="cuda")
-    mean = torch.mean(res_out)
-    std = torch.std(res_out)
-    assert torch.abs(mean) < 0.01
-    assert torch.abs(std - 1) < 0.01
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_rand_like(shape, dtype):
-    x = torch.randn(size=shape, dtype=dtype, device="cuda")
-    with flag_gems.use_gems():
-        res_out = torch.rand_like(x)
-    assert (res_out <= 1.0).all()
-    assert (res_out >= 0.0).all()
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_normal(shape, dtype):
-    loc = torch.full(size=shape, fill_value=3.0, dtype=dtype, device="cuda")
-    scale = torch.full(size=shape, fill_value=10.0, dtype=dtype, device="cuda")
-    with flag_gems.use_gems():
-        res_out = torch.distributions.normal.Normal(loc, scale).sample()
-    mean = torch.mean(res_out)
-    std = torch.std(res_out)
-    assert torch.abs(mean - 3.0) < 0.1
-    assert torch.abs(std - 10.0) < 0.1
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_uniform(shape, dtype):
-    x = torch.randn(size=shape, dtype=dtype, device="cuda")
-    with flag_gems.use_gems():
-        x.uniform_(-3, 3)
-    assert (x <= 3.0).all()
-    assert (x >= -3.0).all()
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
 @pytest.mark.parametrize("dtype", [torch.cfloat])
 def test_accuracy_resolve_neg(shape, dtype):
     x = torch.randn(size=shape, dtype=dtype, device="cuda")
@@ -269,54 +215,3 @@ def test_accuracy_resolve_conj(shape, dtype):
     with flag_gems.use_gems():
         z = y.resolve_conj()
     assert not z.is_conj()
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_zeros(shape, dtype):
-    with flag_gems.use_gems():
-        res_out = torch.zeros(shape, dtype=dtype, device="cuda")
-    gems_assert_equal(res_out, torch.zeros(shape, dtype=dtype, device="cuda"))
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_ones(shape, dtype):
-    with flag_gems.use_gems():
-        res_out = torch.ones(shape, dtype=dtype, device="cuda")
-    gems_assert_equal(res_out, torch.ones(shape, dtype=dtype, device="cuda"))
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_full(shape, dtype):
-    with flag_gems.use_gems():
-        res_out = torch.full(shape, 3.1415926, dtype=dtype, device="cuda")
-    gems_assert_equal(res_out, torch.full(shape, 3.1415926, dtype=dtype, device="cuda"))
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_zeros_like(shape, dtype):
-    x = torch.empty(size=shape, dtype=dtype, device="cuda")
-    with flag_gems.use_gems():
-        res_out = torch.zeros_like(x)
-    gems_assert_equal(res_out, torch.zeros_like(x))
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_ones_like(shape, dtype):
-    x = torch.empty(size=shape, dtype=dtype, device="cuda")
-    with flag_gems.use_gems():
-        res_out = torch.ones_like(x)
-    gems_assert_equal(res_out, torch.ones_like(x))
-
-
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_full_like(shape, dtype):
-    x = torch.empty(size=shape, dtype=dtype, device="cuda")
-    with flag_gems.use_gems():
-        res_out = torch.full_like(x, 3.1415926)
-    gems_assert_equal(res_out, torch.full_like(x, 3.1415926))
